add user handler tests
the changeset also re-introduces the sqlUserStorage, but with multiple tiny
interfaces which only handle the required functionality, e.g. UserTokenUpdater
which only updates the remember_token for a given user.

This allows us to a) extract the SQL from the HTTP handlers, putting less logic
into the handlers, and b) to test the HTTP handlers with the database mocked
away.